+++ src/ast/passes/semantic_analyser.cpp
@@ -2607,7 +2607,9 @@ void SemanticAnalyser::visit(AttachPoint &ap)
       LOG(ERROR, ap.loc, err_) << "kprobes should be attached to a function";
     if (is_final_pass()) {
       // Warn if user tries to attach to a non-traceable function
-      if (!has_wildcard(ap.func) && !bpftrace_.is_traceable_func(ap.func)) {
+      if (bpftrace_.config_.get(ConfigKeyMissingProbes::default_) !=
+              ConfigMissingProbes::ignore &&
+          !has_wildcard(ap.func) && !bpftrace_.is_traceable_func(ap.func)) {
         LOG(WARNING, ap.loc, out_)
             << ap.func
             << " is not traceable (either non-existing, inlined, or marked as "